1202. Smallest String With Swaps
